Revert to [T; 1] representations for mp*_t, add gmp_randstate_ptr
gmp_randstate_t uses a [T; 1] representation, which makes it able
to be decay to a pointer in a function parameter context.

This behavior does not apply in Rust however, not even in FFI contexts,
so this can't be used.

Instead, the functions now take gmp_randstate_ptr as argument.

Also revert a previously attempted, incorrect fix for this problem.